+++ akshare/__init__.py
@@ -2049,9 +2049,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.6.28 add: add currency_boc_safe interface
 1.6.29 fix: fix stock_board_industry_name_em interface
 1.6.30 fix: fix stock_info_a_code_name interface
+1.6.31 fix: fix stock_gdfx_free_holding_detail_em interface
 """
 
-__version__ = "1.6.30"
+__version__ = "1.6.31"
 __author__ = "AKFamily"
 
 import sys