+++ akshare/stock_feature/stock_gdfx_em.py
@@ -10,7 +10,9 @@ import requests
 from tqdm import tqdm
 
 
-def stock_gdfx_free_holding_statistics_em(date: str = "20210630") -> pd.DataFrame:
+def stock_gdfx_free_holding_statistics_em(
+    date: str = "20210630",
+) -> pd.DataFrame:
     """
     东方财富网-数据中心-股东分析-股东持股统计-十大流通股东
     https://data.eastmoney.com/gdfx/HoldingAnalyse.html
@@ -81,15 +83,33 @@ def stock_gdfx_free_holding_statistics_em(date: str = "20210630") -> pd.DataFram
         ]
     ]
     big_df["统计次数"] = pd.to_numeric(big_df["统计次数"])
-    big_df["公告日后涨幅统计-10个交易日-平均涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-10个交易日-平均涨幅"])
-    big_df["公告日后涨幅统计-10个交易日-最大涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-10个交易日-最大涨幅"])
-    big_df["公告日后涨幅统计-10个交易日-最小涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-10个交易日-最小涨幅"])
-    big_df["公告日后涨幅统计-30个交易日-平均涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-30个交易日-平均涨幅"])
-    big_df["公告日后涨幅统计-30个交易日-最大涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-30个交易日-最大涨幅"])
-    big_df["公告日后涨幅统计-30个交易日-最小涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-30个交易日-最小涨幅"])
-    big_df["公告日后涨幅统计-60个交易日-平均涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-60个交易日-平均涨幅"])
-    big_df["公告日后涨幅统计-60个交易日-最大涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-60个交易日-最大涨幅"])
-    big_df["公告日后涨幅统计-60个交易日-最小涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-60个交易日-最小涨幅"])
+    big_df["公告日后涨幅统计-10个交易日-平均涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-10个交易日-平均涨幅"]
+    )
+    big_df["公告日后涨幅统计-10个交易日-最大涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-10个交易日-最大涨幅"]
+    )
+    big_df["公告日后涨幅统计-10个交易日-最小涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-10个交易日-最小涨幅"]
+    )
+    big_df["公告日后涨幅统计-30个交易日-平均涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-30个交易日-平均涨幅"]
+    )
+    big_df["公告日后涨幅统计-30个交易日-最大涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-30个交易日-最大涨幅"]
+    )
+    big_df["公告日后涨幅统计-30个交易日-最小涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-30个交易日-最小涨幅"]
+    )
+    big_df["公告日后涨幅统计-60个交易日-平均涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-60个交易日-平均涨幅"]
+    )
+    big_df["公告日后涨幅统计-60个交易日-最大涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-60个交易日-最大涨幅"]
+    )
+    big_df["公告日后涨幅统计-60个交易日-最小涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-60个交易日-最小涨幅"]
+    )
     return big_df
 
 
@@ -165,15 +185,33 @@ def stock_gdfx_holding_statistics_em(date: str = "20210930") -> pd.DataFrame:
         ]
     ]
     big_df["统计次数"] = pd.to_numeric(big_df["统计次数"])
-    big_df["公告日后涨幅统计-10个交易日-平均涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-10个交易日-平均涨幅"])
-    big_df["公告日后涨幅统计-10个交易日-最大涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-10个交易日-最大涨幅"])
-    big_df["公告日后涨幅统计-10个交易日-最小涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-10个交易日-最小涨幅"])
-    big_df["公告日后涨幅统计-30个交易日-平均涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-30个交易日-平均涨幅"])
-    big_df["公告日后涨幅统计-30个交易日-最大涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-30个交易日-最大涨幅"])
-    big_df["公告日后涨幅统计-30个交易日-最小涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-30个交易日-最小涨幅"])
-    big_df["公告日后涨幅统计-60个交易日-平均涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-60个交易日-平均涨幅"])
-    big_df["公告日后涨幅统计-60个交易日-最大涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-60个交易日-最大涨幅"])
-    big_df["公告日后涨幅统计-60个交易日-最小涨幅"] = pd.to_numeric(big_df["公告日后涨幅统计-60个交易日-最小涨幅"])
+    big_df["公告日后涨幅统计-10个交易日-平均涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-10个交易日-平均涨幅"]
+    )
+    big_df["公告日后涨幅统计-10个交易日-最大涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-10个交易日-最大涨幅"]
+    )
+    big_df["公告日后涨幅统计-10个交易日-最小涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-10个交易日-最小涨幅"]
+    )
+    big_df["公告日后涨幅统计-30个交易日-平均涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-30个交易日-平均涨幅"]
+    )
+    big_df["公告日后涨幅统计-30个交易日-最大涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-30个交易日-最大涨幅"]
+    )
+    big_df["公告日后涨幅统计-30个交易日-最小涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-30个交易日-最小涨幅"]
+    )
+    big_df["公告日后涨幅统计-60个交易日-平均涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-60个交易日-平均涨幅"]
+    )
+    big_df["公告日后涨幅统计-60个交易日-最大涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-60个交易日-最大涨幅"]
+    )
+    big_df["公告日后涨幅统计-60个交易日-最小涨幅"] = pd.to_numeric(
+        big_df["公告日后涨幅统计-60个交易日-最小涨幅"]
+    )
     return big_df
 
 
@@ -344,9 +382,7 @@ def stock_gdfx_free_top_10_em(
     :return: 十大股东
     :rtype: pandas.DataFrame
     """
-    url = (
-        "https://emweb.securities.eastmoney.com/PC_HSF10/ShareholderResearch/PageSDLTGD"
-    )
+    url = "https://emweb.securities.eastmoney.com/PC_HSF10/ShareholderResearch/PageSDLTGD"
     params = {
         "code": f"{symbol.upper()}",
         "date": f"{'-'.join([date[:4], date[4:6], date[6:]])}",
@@ -483,11 +519,11 @@ def stock_gdfx_free_holding_detail_em(date: str = "20210930") -> pd.DataFrame:
         "报告期",
         "股东名称",
         "期末持股-数量",
-        "期末持股-持股占流通股比",
+        "-",
+        "期末持股-持股变动",
         "-",
         "-",
         "-",
-        "股东排名",
         "股票简称",
         "-",
         "-",
@@ -496,7 +532,7 @@ def stock_gdfx_free_holding_detail_em(date: str = "20210930") -> pd.DataFrame:
         "-",
         "-",
         "期末持股-数量变化比例",
-        "-",
+        "股东类型",
         "-",
         "公告日",
         "-",
@@ -505,25 +541,24 @@ def stock_gdfx_free_holding_detail_em(date: str = "20210930") -> pd.DataFrame:
         "-",
         "-",
         "-",
-        "股东类型",
-        "期末持股-持股变动",
+        "-",
+        "-",
         "-",
         "-",
         "-",
         "-",
         "期末持股-数量变化",
+        "-",
     ]
     big_df = big_df[
         [
             "序号",
             "股东名称",
             "股东类型",
-            "股东排名",
             "股票代码",
             "股票简称",
             "报告期",
             "期末持股-数量",
-            "期末持股-持股占流通股比",
             "期末持股-数量变化",
             "期末持股-数量变化比例",
             "期末持股-持股变动",
@@ -534,7 +569,6 @@ def stock_gdfx_free_holding_detail_em(date: str = "20210930") -> pd.DataFrame:
     big_df["报告期"] = pd.to_datetime(big_df["报告期"]).dt.date
     big_df["公告日"] = pd.to_datetime(big_df["公告日"]).dt.date
     big_df["期末持股-数量"] = pd.to_numeric(big_df["期末持股-数量"])
-    big_df["期末持股-持股占流通股比"] = pd.to_numeric(big_df["期末持股-持股占流通股比"])
     big_df["期末持股-数量变化"] = pd.to_numeric(big_df["期末持股-数量变化"])
     big_df["期末持股-数量变化比例"] = pd.to_numeric(big_df["期末持股-数量变化比例"])
     big_df["期末持股-流通市值"] = pd.to_numeric(big_df["期末持股-流通市值"])
@@ -956,10 +990,14 @@ def stock_gdfx_holding_teamwork_em() -> pd.DataFrame:
 
 
 if __name__ == "__main__":
-    stock_gdfx_free_holding_statistics_em_df = stock_gdfx_free_holding_statistics_em(date="20210930")
+    stock_gdfx_free_holding_statistics_em_df = (
+        stock_gdfx_free_holding_statistics_em(date="20210930")
+    )
     print(stock_gdfx_free_holding_statistics_em_df)
 
-    stock_gdfx_holding_statistics_em_df = stock_gdfx_holding_statistics_em(date="20210930")
+    stock_gdfx_holding_statistics_em_df = stock_gdfx_holding_statistics_em(
+        date="20210930"
+    )
     print(stock_gdfx_holding_statistics_em_df)
 
     stock_gdfx_free_holding_change_em_df = stock_gdfx_free_holding_change_em(
@@ -967,7 +1005,9 @@ if __name__ == "__main__":
     )
     print(stock_gdfx_free_holding_change_em_df)
 
-    stock_gdfx_holding_change_em_df = stock_gdfx_holding_change_em(date="20210930")
+    stock_gdfx_holding_change_em_df = stock_gdfx_holding_change_em(
+        date="20210930"
+    )
     print(stock_gdfx_holding_change_em_df)
 
     stock_gdfx_free_top_10_em_df = stock_gdfx_free_top_10_em(
@@ -975,7 +1015,9 @@ if __name__ == "__main__":
     )
     print(stock_gdfx_free_top_10_em_df)
 
-    stock_gdfx_top_10_em_df = stock_gdfx_top_10_em(symbol="sh688686", date="20210630")
+    stock_gdfx_top_10_em_df = stock_gdfx_top_10_em(
+        symbol="sh688686", date="20210630"
+    )
     print(stock_gdfx_top_10_em_df)
 
     stock_gdfx_free_holding_detail_em_df = stock_gdfx_free_holding_detail_em(
@@ -983,7 +1025,9 @@ if __name__ == "__main__":
     )
     print(stock_gdfx_free_holding_detail_em_df)
 
-    stock_gdfx_holding_detail_em_df = stock_gdfx_holding_detail_em(date="20210930")
+    stock_gdfx_holding_detail_em_df = stock_gdfx_holding_detail_em(
+        date="20210930"
+    )
     print(stock_gdfx_holding_detail_em_df)
 
     stock_gdfx_free_holding_analyse_em_df = stock_gdfx_free_holding_analyse_em(
@@ -991,10 +1035,14 @@ if __name__ == "__main__":
     )
     print(stock_gdfx_free_holding_analyse_em_df)
 
-    stock_gdfx_holding_analyse_em_df = stock_gdfx_holding_analyse_em(date="20210930")
+    stock_gdfx_holding_analyse_em_df = stock_gdfx_holding_analyse_em(
+        date="20210930"
+    )
     print(stock_gdfx_holding_analyse_em_df)
 
-    stock_gdfx_free_holding_teamwork_em_df = stock_gdfx_free_holding_teamwork_em()
+    stock_gdfx_free_holding_teamwork_em_df = (
+        stock_gdfx_free_holding_teamwork_em()
+    )
     print(stock_gdfx_free_holding_teamwork_em_df)
 
     stock_gdfx_holding_teamwork_em_df = stock_gdfx_holding_teamwork_em()